@@ -1,28 +1,24 @@
-'use strict';
+export default async function pReduce(iterable, reducer, initialValue) {
+	return new Promise((resolve, reject) => {
+		const iterator = iterable[Symbol.iterator]();
+		let index = 0;
 
-const pReduce = (iterable, reducer, initialValue) => new Promise((resolve, reject) => {
-	const iterator = iterable[Symbol.iterator]();
-	let index = 0;
+		const next = async total => {
+			const element = iterator.next();
 
-	const next = async total => {
-		const element = iterator.next();
+			if (element.done) {
+				resolve(total);
+				return;
+			}
 
-		if (element.done) {
-			resolve(total);
-			return;
-		}
+			try {
+				const [resolvedTotal, resolvedValue] = await Promise.all([total, element.value]);
+				next(reducer(resolvedTotal, resolvedValue, index++));
+			} catch (error) {
+				reject(error);
+			}
+		};
 
-		try {
-			const value = await Promise.all([total, element.value]);
-			next(reducer(value[0], value[1], index++));
-		} catch (error) {
-			reject(error);
-		}
-	};
-
-	next(initialValue);
-});
-
-module.exports = pReduce;
-// TODO: Remove this for the next major release
-module.exports.default = pReduce;
+		next(initialValue);
+	});
+}
